Armor and Damage Queries

+++ src/Sidekick.Business/Filters/FilterValue.cs
@@ -2,7 +2,7 @@ namespace Sidekick.Business.Filters
 {
     public class FilterValue
     {
-        public int? Min { get; set; }
-        public int? Max { get; set; }
+        public double? Min { get; set; }
+        public double? Max { get; set; }
     }
 }
